requirements.txt
python-dotenv>=0.19.0
PyYAML>=6.0
requests>=2.26.0
flask>=2.0.1
gunicorn>=20.1.0
prometheus-client>=0.11.0
SQLAlchemy>=2.0.0
passlib[bcrypt]>=1.7.4 